requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
drf-nested-routers==0.94.1
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
